Refactored file names and code formatting
