+++ TutoringApp/components/atoms/tutor.jsx
@@ -15,11 +15,12 @@ function getRatingEmoji(rating) {
 }
 
 const Tutor = ({ name, course, rating }) => (
-  <Card style={{
-    borderRadius: 15,
-    marginVertical: "0.25%",
-    
-  }}>
+  <Card
+    style={{
+      borderRadius: 15,
+      marginVertical: "0.25%",
+    }}
+  >
     <Card.Content>
       <Text>{name}</Text>
       <Text>{course}</Text>
